requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
begins==0.9
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cssselect2==0.6.0
Cython==0.29.24
decorator==5.1.0
docx==0.2.4
docxcompose==1.3.3
docxtpl==0.14.2
easycython==1.0.7
fonttools==4.33.3
future==0.18.2
geocoder==1.38.1
html5lib==1.1
idna==3.3
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.3
pandas==1.3.4
peewee==3.14.8
Pillow==8.4.0
pyasn1==0.4.8
pycparser==2.21
pydyf==0.1.2
PyMySQL==1.0.2
PyPDF3==1.0.5
pyphen==0.12.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
ratelim==0.1.6
requests==2.26.0
rsa==4.8
sentry-sdk==1.5.10
six==1.16.0
tinycss2==1.1.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
weasyprint==54.3
webencodings==0.5.1
yarl==1.7.0
zopfli==0.2.1
